+++ src/routes/ManageAds/AdsContainers/adsContainers.module.scss
@@ -1,8 +1,9 @@
 @use 'styles/mixins/flexbox';
 
 .adsContainersWrapper {
-  @include flexbox.flexbox(between, center);
+  @include flexbox.flexbox(start, center);
   flex-wrap: wrap;
   width: 100%;
   height: 100%;
+  overflow-x: scroll;
 }
