--- src/routes/ManageAds/AdsHeader/AdsCreateButton/adsCreateButton.module.scss
@@ -1,14 +1,0 @@
-@use 'styles/constants/colors';
-@use 'styles/mixins/flexbox';
-
-.createWrapper {
-  position: relative;
-}
-
-.createButton {
-  width: 100px;
-  height: 40px;
-  color: colors.$WHITE;
-  background-color: colors.$PRIMARY;
-  border-radius: 10px;
-}
