+++ src/routes/ManageAds/AdsHeader/adsHeader.module.scss
@@ -1,6 +1,17 @@
+@use 'styles/constants/colors';
 @use 'styles/mixins/flexbox';
 
 .adsHeaderWrapper {
   @include flexbox.flexbox(between, center);
   padding-bottom: 30px;
+  overflow-x: scroll;
+}
+
+.createButton {
+  width: 100px;
+  min-width: 100px;
+  height: 40px;
+  color: colors.$WHITE;
+  background-color: colors.$PRIMARY;
+  border-radius: 10px;
 }
